requirements.txt
websocket-client
requests
pandas
pyupbit
discord.py